requirements.txt
cython
numpy
asgiref==3.4.1
Babel==2.9.1
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Django==3.2.*
django-debug-toolbar==3.2.2
django-mysql==4.0.0
django-toast-messages==0.1
djangorestframework==3.12.4
elasticsearch==7.14.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
filelock==3.0.12
google-api-core==1.31.1
google-api-python-client==2.21.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6 
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
isodate==0.6.0
jdcal==1.4.1
Jinja2==3.0.1
joblib==1.0.1
langdetect==1.0.9
Markdown==3.3.4
MarkupSafe==2.0.1
murmurhash==1.0.5
mysqlclient==2.0.3
oauth2client==4.1.3
opencv-python==4.5.3.56
openpyxl==3.0.*
packaging==21.0
pandas==1.3.*
pathy==0.6.0
Pillow==8.3.2
plum-dispatch==1.5.4
preshed==3.0.5
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pycco==0.6.0
pydantic==1.8.2
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pystache==0.6.*
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-file==1.5.1
rsa==4.7.2
six==1.16.0
smart-open==5.2.1
smartypants==2.0.1
sqlparse==0.4.2
srsly==2.4.1
thinc==8.0.10
tldextract==3.1.2
tqdm
typer==0.3.2
uritemplate==3.0.1
urllib3==1.26.6
wasabi==0.8.2
XlsxWriter==1.4.5
